requirements.txt
pandas
streamlit
prophet>=1.1
statsmodels
openpyxl
matplotlib-venn
numpy
plotly 
scikit-learn
